feat: move tracks to track
