+++ src/app.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { UserModule } from './user/user.module';
-import { TracksModule } from './tracks/tracks.module';
+import { TracksModule } from './track/track.module';
 import { ArtistsModule } from './artists/artists.module';
 
 @Module({
